8/13/18
I realize the if statement is redundant but I'm still having issues and am not quite ready to get rid of it... and I'm lazy. The main issue is at the finish of the for loop, FIJI has some sort of background code which wraps all the edited photos together and re-processes them. It's a minor inconvenience but it's really annoying that this isn't a hands off macro. If used, expect to hit cancel when it asks to convert the stack to binary, then save the summary table by hand.