Create H_AEC_High_Threshold
Due to issues getting the computer to differentiate between RBC and AEC positive, the threshold was increased in attempt to get the computer to distinguish the two. It's not perfect so if you get inconsistent numbers on a sample, look at those pictures by hand to see what the computer was counting as positive. 
Occasionally, the first picture in the folder will not have any positive cells and will throw an error "Nothing in the ROI manager"; this refers to the circling of the positive cells on which it can't find anything in the ROI manager so it can't circle and aborts. In this case, comment out (//) lines 39-42 and re-run.